gplazma: fix embedded gPlazma
gPlazma may be run 'embedded'.  This is actually nothing more than 
short-hand for running a gPlazma cell in the same domain as a door, 
yet we preserve the nomenclature to be backwards compatible with older 
configuration.  A recent update removed gPlazma v1, leaving only 
gPlazma v2.  When this happened, the gplazma.version property was 
dropped and the Spring XML filename was modified.

Unfortunately, the embedded gPlazma batch fragment was mistakenly 
skipped, resulting in breaking embedded gPlazma.

This patch fixes this.

Target: trunk
Patch: http://rb.dcache.org/r/4264
Acked-by: Tigran Mkrtchyan