srm: use spring bean profiles
The SRM may be configured to use a synchronous or asynchronous 
connector.  Currently this choice is reflected by splitting the spring 
XML file into several parts and selecting the filename at runtime based 
on the configured type.

This patch adjusts SRM so it uses bean profiles to choose which 
connector is in use.

Target: trunk
Patch: http://rb.dcache.org/r/4261
Acked-by: Tigran Mkrtchyan