webadmin: remove dependency on gplazma.version property
Patch 17035 removed gPlazma-1 and, in doing so, marked the 
'gplazma.version' property as obsolete, so removing it from the 
properties available to a batch script.

Unfortunately, the webadmin service still depended on 'gplazma.version' 
as it had two mechanisms with which it tried to login in a user.  The 
consequence was that, unless the admin configures the 'gplazma.version' 
property, the webadmin service would cause its hosting domain to fail 
on startup.  The ironic part is that, when 'gplazma.version' is 
defined, dCache would warn that the property is obsolete and should be 
deleted yet, by deleting this property, the domain would fail.

This patch removes the outdated gPlazma-1 code from webadmin and, by 
doing so, allow webadmin to start.

Target: trunk
Patch: http://rb.dcache.org/r/4262
Acked-by: Tigran Mkrtchyan