Modified IPMatcher to be more flexible in terms of a users environment.
It turned out, that Mac OS X apparently appends it's default domain in some cases even to localhost, which leads to failing of the testMixed Test within IPMatcherTest.java. To fix that, I created a Nested Class which is responsible for DNS lookups, which in turn is being stubbed within the test to return a predictable output.

Patch: http://rb.dcache.org/r/4249/
Acked-by: Karsten Schwank
Target: trunk