gplazma: don't add "Login failed: " to message
gPlazma currently adds "Login failed: " as a prefix to the specific 
message coming from the plugin that failed the login.

The is redundant since the door that requested the login will add this 
information, usually as a door-specific prefix.  The result is the 
message reported to the end-user contains unnecessary duplication of 
information (see Testing Done section below for examples)

This patch simply removes "Login failed: " part as only the door knows 
precisely what operation is failing and will add this information 
anyway.

Target: trunk
Patch: http://rb.dcache.org/r/4273/
Acked-by: Karsten Schwank