gplazma: refactor unit tests based on BDD
I was adding some (small) additional functionality to gPlazma and 
wanted to add some unit-tests for them.  The current state of the 
unit-tests were so awful that I rewrote them.  I've attempted to 
follow Gerd's suggested BDD-style; albeit in a limit fashion.

This patch refactors only the DefaultAccountStrategy tests but 
provides the blueprint for refactoring the others.  If people are 
happy with this style then I'll fix the other tests, too.

Target: trunk
Patch: http://rb.dcache.org/r/4256/
Acked-by: Karsten Schwank