gplazma: catch plugin bugs
gPlazma includes support for plugins; these plugins may contain bugs.  
Currently, if a plugin throws RuntimeException then this exception will 
propagate through gPlazma and be caught in the dCache-gPlazma interface 
layer and recorded as a gPlazma bug.

This is undesirable because the error message is misleading, 
(potentially) information is lost and the failure may be irrelevant to 
a login attempt.

This patch adds support for catching plugin bugs within gPlazma.  If a 
bug is found then it is logged with the corresponding stack-trace.

A plugin where a bug is triggered is treated as if the plugin threw 
AuthenticationException.  For plugins used during the login process, 
whether this matters depends on how the plugin is configured: if a 
REQUISITE plugin throws a RuntimeException then the login will fail, 
however if the buggy plugin is configured as OPTIONAL then the login 
will proceed.

Note that the login may still fail even if the plugin is configured as 
OPTIONAL if, for this particular user, the faulty plugin's result is 
needed.

Target: trunk
Patch: http://rb.dcache.org/r/4275
Acked-by: Tigran Mkrtchyan