cells/gPlazma: fix race condition when killing a cell
gPlazma may be run "as a module" within a door.  The term "module" is 
now a misnomer: it actually runs as a normal cell, but one that is 
configured not to register itself as a well-known cell.  The door will 
use the domain-local gPlazma preferentially over the central one.  
However, an admin may also (mistakenly) configure the central gPlazma 
to run in the same domain.  Clearly there cannot be two cells with the 
same name; therefore the gplazma service script kills the "as a module" 
gPlazma instance to allow the central gPlazma cell to run.  To achieve 
this, the script calls the kill command, which instructs the cell to 
terminate.

Unfortunately, although the kill admin command triggers the cell to 
die, it doesn't guarantee that the cell has finished shutting down 
before returning.  The effect is a race between the "as a module" 
gPlazma cell terminating and the gplazma service batch file starting up 
the central gPlazma cell.  The call to create the central gPlazma cell 
"wins", resulting in an attempt to create a new gPlazma cell while the 
"as a module" gPlazma cell still exists.  This attempt fails, resulting 
in the entire domain shutting down.

This patch fixes the problem by guaranteeing the kill command doesn't 
return until the cell has finished shutting down.

The patch also updates the error message, which didn't adequately 
explain what is happening or what the admin should do to fix the 
problem.

Target: trunk
Request: 2.1
Request: 1.9.12
Patch: http://rb.dcache.org/r/4278
Acked-by: Karsten Schwank