gplazma: hid reason why login failed from end-user
A common approach to security is to avoid specifying why a login 
failed; in particular, we want to avoid giving any clue whether a 
particular username (DN, Kerberos principal, ...) is known to the 
system.  This is to prevent someone trying to discover which users are 
known to the system.

Target: trunk
Patch: http://rb.dcache.org/r/4286
Acked-by: Karsten Schwank