core: fix net unit matching for world-net (0.0.0.0/0) subnet
Although IPv4 addresses are 32-bit wide, we need to use Java's 64-bit 
integer value (long) since we want to store the numbers as unsigned 
integers values: [0, 4294967295] and all Java primative types are 
signed.

NetHandler handles the conversion to a long value with code like:

    (new BigInteger(address.getAddress()).longValue();

This doesn't give the desired result since BigInteger treats the 32-bit 
number from the IPv4 address as a signed 32-bit number.  The longValue() 
promotes this signed 32-bit value to a signed 64-bit value.  This means 
than any IPv4 address where the first octet is >= 128 will have a 
negative long value; i.e., the most significant 32 bits all '1', like 
"0xffffffff--------".

The added subnet is stored against these potentially negative numbers.

NetHandler also allows searching for a subnet that a given IP address is 
within.  When searching for an IPv4 address, the same algorithm is used 
to convert the address to a long: the same mapping artifact is present 
(first octet >= 128 means number is negative).The masks used to filter 
an address (to test membership of a subnet) uses a progressive form: 
initially 0xffffffffffffffff, 0xfffffffffffffffe, 0xfffffffffffffffc and 
so on.  Therefore, these masks allow the high-order 32-bit, allowing 
these numbers to continue to be negative.

Therefore the code works for IPv4 addresses with first octets >= 128, 
but it's a "works by accident".

The problem comes with subnets like 0.0.0.0/0.  Since the highest order 
bit shouldn't be significant (this subnet should match all IPv4 
addresses), the subnet should match IPv4 with first octet < 128 
(addresses that map to a positive long) and those >= 128 (addresses that 
map to a negative long).

However, for a 'a.b.c.d/0' subnet is always stored as 0: i.e, the high 
32 bits are all cleared.  Since the mask used when testing 
subnet-membership for an IPv4 address includes the high 32-bits.  This 
means that 0.0.0.0/0 matches only IPv4 address with the first octet <= 
127 since only these addresses will have the high 32-bits cleared.

Note that our default configuration includes a unit-group 'world-net', 
that uses the '0.0.0.0/0.0.0.0' net unit.  Because of this bug, the 
default configuration is broken for IPv4 addresses with first octet >= 
128; i.e., trunk is currently broken for default configuration and 
probably vast majority of deployed dCaches.

This patch fixes the problem by introducing a new method for converting 
an IPv4 address to a long value.

Target: trunk
Patch: http://rb.dcache.org/r/4305/
Acked-by: Karsten Schwank