build: Add relocatable tar packaging
The existing tar package for the FHS layout is practically useless. It
would only work if unpacked in the root of the directory hierarchy and
it would add files in locations usually reserved for package managers.

This patch adds a relocatable tar package. It is relocatable in the
sense that in can be unpacked anywhere and run from that location.
dCache will not write any files outside the installation directory,
meaning configuration files, log files, etc are all stored in the
installation directory.

This layout is perfect for somebody who just wants to try dCache. Unless
a service on a priviledged port is needed, dCache can run as an
unpriviledged user.

The package may also serve as an alternative for those users that want
to keep installing in /opt. Such users could simply unpack this tarball
in that location and use it directly.

In its current form the package is hard to upgrade as one would manually
have to copy the etc and var directories from the old to the new
install. We may be able to solve that problem by letting
/etc/defaults/dcache contain a DCACHE_SETUP variable pointing to an
alternative dcache.conf. That would allow all etc and var locations to
be redefined and would mean one could easily unpack a new version of the
tarball and keep the old configuration.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4105/