poolmanager: Make default partition type configurable
The default partition was hardcoded to the type 'classic'. This patch
allows the type of the default partition to be changed.


The approach taken is to allow partitions to be recreated with a
different type. The existing partition is overwritten. This allows the
default partition to be recreated with a different type.


One complication was that we for reasons of backwards compatibility
didn't distinguish between the set of inherited properties and the
properties of the default partition. In 1.9.12 there was no need to
distinguish because all partitions were of the same type and thus
accepted the same set of parameters. Now that several partition types
exist, inheriting from the default partition doesn't make sense, which
is why we distinguish between the inherited properties and the
properties of the default partition. To provide compatibility with the
old scheme, setting a property on the default partition was silently
treated as an update to the inherited properties. This has now been
changed such that the properties of the default partition and the
inherited properties can be updated individually. This change however
only applies if a default partition was explicitly created - otherwise
an update to the default partition's properties is still applied to the
inherited set. This ensure backwards compatibility with 1.9.12. This
feature can be removed once we move to dCache 2.3.


The default setup embedded into the service definition batch file is
updated to create an explicit default partition. Thus a fresh
installation will always use an explicitly defined default.

The patch updates the name of the _properties field to better reflect
the meaning of this set.

The patch also adds help text to all PartitionManager commands.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4212/