poolmanager: Change default partition type to wass
This patch changes the default type of the default partition to wass.
This change only applies to fresh installations without a customized
pool manager setup. Installations that already have a poolmanager.conf
file will continue to use 'classic' as the type of the default
partition.


We cannot yet change the default type used when creating new partitions
with 'pm create'. This is because a 1.9.12 poolmanager.conf file does
not provide a partition type. Thus we have to default to classic. In
dCache 2.3 we can change the default to wass (or make the type a
mandatory parameter). This is because dCache 2.0 and newer generate
poolmanager.conf files with an explicit partition type.

Release notes: Suggest people regenerate poolmanager.conf after upgrade 
by running the save command in pool manager. This will ensure 
compatibility with future releases.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4213/