common: clarify user-name principals
dCache common contains two class that implement Principal that define 
some form of simple name for the user: LoginNamePrincipal and 
UserNamePrincipal.  The former is for the requested username while the 
other is for the username identified as part of the login process.

The current code-base has JavaDoc that is, at best, confusing.  This 
patch updates this to make the role that these principals have clearer.

Target: trunk
Patch: http://rb.dcache.org/r/4316
Acked-by: Tigran Mkrtchyan