gplazma: add login monitoring and admin command that uses it
Currently, gPlazma behaviour is like a black-box: when a login 
request comes in there's no mechanism to discover how the request 
propagates through the plugins.  This lack of information makes it 
hard to understand why a particular login request succeeds or fails, 
or why certain principals are available or not.

This patch adds the concept of a LoginMonitor: a class that 
implements this interface receives information on how a login request 
is processed.  The effect of the plugins and phases are provided 
along with whether the plugins and phases succeed or fail.  If a 
plugin fails then the error message is provided.  The plugin control 
(REQUIRED / REQUISITE/ OPTIONAL / SUFFICIENT) is also shown, along 
with the aggregated result (OPTIONAL + FAIL => OK).

Two LoginMonitor implementations are provided: one that logs data at 
debug level (replacing current hard-coded behavior) and one that 
records the details in a compact class LoginResult.

The patch includes a class that provides an ASCII-art view of a 
LoginResult.

The patch provides two new admin commands for gPlazma.  The "test 
login" command replaces the existing "get mapping" command but is 
more flexible.  The new "explain login" command uses the same 
arguments as "test login" but provides an ASCII-art explaination of 
the login attempt.

The current patch is limited to looking at only principals and only 
for the login command.  Anticipated extension is to add support for 
credentials and for the two identity operations.  Some of the plugin 
error messages should be tweaked.  We may want to produce ASCII-art 
explanations in the log files (if careful to rate-limited).

FOR RELEASE NOTES:

This patch adds two new commands to the admin interface when looking 
at the gPlazma cell.  "test login" replaces the existing "get 
mapping" command.  It shows the result of a login attempt but is more 
flexible when describing which principals have been identified for 
the user.

The second is "explain login".  This command uses the same format as 
"test login" but provides detailed information on how the login was 
processed.  The result of each processed plugin is explained in a 
tree structure.

Target: trunk
Patch: http://rb.dcache.org/r/4312