srm: always listen for SSL connections
One of the evolving projects is SRM-over-SSL.  This is the same SRM 
protocol(s) but using an SSL transport rather than a GSI one.  There 
are several advantages of this, mostly centred on SSL being an 
industry standard.  For some time, the grid will likely need both GSI 
and SSL based SRM.  This is to allow a smooth transition between the 
two transports.

Unfortunately, GSI and SSL are incompatible and a server cannot 
easily or reliably detect whether an end-client is GSI or SSL.  
Therefore, to support both transports, the SRM must listen on 
different ports.

Currently dCache has support for choosing whether an SRM instance 
will listen on a GSI transport or an SSL one.  While this is OK for 
testing, it doesn't work for production environments because we don't 
currently support multiple SRMs running in the same dCache instance.

This patch removes the SRM option of whether to listen using a GSI or 
SSL transport.  Instead, an SRM will listen on two ports: one for GSI 
and the other for SSL.  This has the advantage of being easier to 
configure and sites need to run only a single SRM instance to support 
both.

RELEASE NOTES:

The srm service now listens on two ports.  These are, by default, 
8443 (as before) and 8445.  Port 8443 continues to be for SRM clients 
that use GSI-based communication and the new port is for SRM clients 
that use SSL.  While the SSL port is configurable, it is recommended 
to use the default as this an agreed port-number for SSL-based SRM 
traffic.

Target: trunk
Patch: http://rb.dcache.org/r/4276/
Acked-by: Tigran Mkrtchyan