Gplazma2-Argus: Integration Tests
The tests for the Argus plugins partly depend on a real Argus instance. Those tests are not 
really Unit-Tests, but rather integration tests. 

This patch divides the tests into unit tests and integration tests. The latter are executed
by the maven surefire plugin (added in patch http://rb.dcache.org/r/4325/) in the verify step. 
The integration tests depend on a running Argus instance on swords.desy.de, which is accessible
from outside DESY and is preconfigured as described in the integration tests file.

Patch: http://rb.dcache.org/r/4331/
Acked-by: Paul Millar,
Target: default