configuration: add 'one-of' annotation
dCache has several configuration properties that take one of only a few 
possible valid values.  Such properties are closed enumerations.

We currently have no simple, overarching way of specifying enumerations 
or to validate such property's value.  Instead, each property requires 
independent verification.  With spring, this may lead to 
incomprehensable error messages as beans failing to instantiate or 
spring itself attempting to read a non-existent XML file.

Worse, an unacceptable value may trigger no error message but dCache 
adopting (apparently) the wrong behaviour.  An example of this is any 
boolean property configured through spring.  If such a property has a 
value other than "true" (e.g., "yes" or some misspelling of "true") is 
has value false.  The same problem is present if code uses the 
Boolean.valueOf method to parse a property's value.

This patch introduces the 'one-of' annotation.  It provides a 
systematic way of describing that a property's value is only one of a 
set of possible values.  If the administrator attempts to configure 
dCache with a value outside that range then an error message is logged 
and dCache refuses to start.

The 'one-of' annotation is similar to other annotations (such as 
"obsolete", "not-for-services") by appearing before the property's name 
and inside parentheses; however, unlike the other annotations, 'one-of' 
is parameterised.  The '?' character distinguishes between the end of 
the annotation label and the start of the parameter.  Other characters 
(like ':' and '=') already have significance so cannot be used.  For 
the 'one-of' annotation, the parameter is a list of possible values, 
separated by the '|' character.

For example, the property 'some.property' that accepts three possible 
values 'foo', 'bar' and 'baz' (with 'bar' the default) may be written:

    (one-of?foo|bar|baz)some.property = bar

Note that the default value must be a valid choice.  If this isn't the 
case then dCache will emit an error message and refuse to start, even 
if administrator-supplied configuration corrects this.

When writing a 'one-of' annotation, the annotation may become fairly 
long.  Judicious use of the logical-line escaping (i.e., '\') allows 
such properties to be written in an approchable fashion; e.g.:

    (deprecated,one-of?very-very-long-possible-value\
                      |another-very-long-possible-value\
                      |yet-another-long-value)\
    some.property = another-very-long-possible-value

Note that, in this case, no white-space may appear between the possible 
value and the '\', and between the '|' separator and the next value; 
however, white-space may be added between '\' and the start of the next 
line as such space is ignored.

This patch also adds appropriate 'one-of' annotations to dCache 
properties.  The property documentation that describes what are valid 
values is removed as it's now redundant.

TODO: add possibility to mark some of an enumerated property's possible 
values as deprecated.  This would allow us to remove support for some 
options in a systematic fashion.

FOR RELEASE NOTES:

Some of dCache properties accept only a few possible values (e.g., 
either "true" or "false").  The parsing of such properties is now more 
robust and incorrect values will be reported.  Previously, incorrect 
values would take some default value.  Now, if a property is configured 
with a value outside the set of acceptable values then an error is 
reported and dCache refuses to start.

As part of this process, some equivalent property values are no longer 
considered the same; for example, some properties took "true", "True" 
and "TRUE" as values that would result in the same behaviour but now 
only "true" is valid.

After upgrading, we recommend you run the "dcache check-config" command 
to check your configuration.

Target: trunk