sshv2: don't ignore NullPointerExceptions
If a user issues a command through the SSH v2 interface that results in 
an exception where getMessage() returns null then this exception is 
silently ignored.  NPE is an example of such an Exception, but others 
may exist.

This patch does two things: if the command triggers a RuntimeException 
(that hasn't been caught as a special case) then a clear error message 
is written to the terminal describing the problem as a bug and 
instructing the user how they can help fix the problem (i.e., send a 
report to us with the information).

For non-RuntimeException, the patch makes sure that the result is never 
null; so something is always returned to the user.

Target: trunk
Request: 2.1
Acked-by: Karsten Schwank
Patch: http://rb.dcache.org/r/4343/