gplazma: log an explanation of why a login failed
This patch adds support for logging why a user was unable to log into 
dCache.  It uses the same mechanism as the "explain login" to show the 
result of gPlazma-2, but the output is stored in the dCache log file.

This patch includes support for suppressing multiple explanations if 
the same user attempts to login multiple times.  This is to prevent a 
flood of login attempts from flooding the log file with explanations.

Target: trunk
Patch: http://rb.dcache.org/r/4346
Acked-by: Tigran Mkrtchyan