WebAdmin: change Locale to use the US format
The web-admin page showing pool usage depends on the locale that the
dCache domain is running under. The generated HTML includes some width
information for the bars showing pool usage. These must have
floating-point numbers using a point for the decimal (e.g., "4.3") but
if a German locale is in use then the floating point numbers are
represented with a comma (e.g., "4,3"). This is an illegal format and
ignored by web-browsers. 

This patch changes the Locale to the US format.

Ticket: #7058, #7064
Acked-by: Paul Millar
Patch: https://rb.dcache.org/r/4355/