Make webadmin/JettyCell configuration more flexible.
In the UniversalSpringCell there is a UniversalSpringCellApplicationContext which makes all the command-line arguments available as properties.

In the webadmin module, however, there is a JettyCell which deals with its command-line arguments by assigning them to internal option fields; the webadmin application implementation proper then uses what it calls a ConfigHolder which, upon instantiation, pulls the JettyCell out of javax.naming.InitialContext and copies to itself the previously mentioned field values; these are then accessed via getters in the WebAdminInterface.xml file as #{ConfigHolder.<field>}.  

The reason for this difference is because dCache in the latter case does not directly control the Spring application context (Wicket does).

The problem here is one of configurability.  To add new wicket pages requiring additional properties, one must not only modify the webadmin.properties and webadmin.batch files, but also the JettyCell and ConfigHolder java classes.  

This patch proposes a less invasive way of providing web-app settings (and JettyCell/dCache objects) to the Wicket spring application context.  Aside from options that are actually internally used by the JettyCell, all command-line arguments are set as System properties by the ConfigHolder, which provides now both the JettyCell and its thread context, as well as all its command-line arguments, to the PropertyPlaceholderConfigurer for that spring file.  Note that the latter must express an explicit depends-on against the ConfigHolder to guarantee the System properties are set before the placeholders are resolved.

With this change, no further alteration of the Java classes is necessary; the properties merely need to be defined and set in the batch file.

Patch: http://rb.dcache.org/r/4328
Ack'd by: Dmitry
Target: trunk

RELEASE NOTES:  

These changes affect only the future webadmin module developers.