build: Rearrange skel for FHS
Rearranges the skel to more closely follow the FHS layout. To be
precise, it follows the layout of the tar package, while the FHS build
reorganizes the structure slightly.

Files that were only included in the old opt packages are dropped. In
particular dCacheConfigure is dropped.


The tar package now includes the var directories. The doc and man
directory have been taken out of the share directory. In contrast to
the other share directories, the primary target of the files in doc
and man is the admin and they should be made visible (in the FHS
layout the same two diretories are placed in the appropriate man
and doc directory).

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4417/