rpc: use local interface to bind and access portmap.
Portmap usually only allows set and unset calls over the loopback
interface. Chimera currently doesn't use the loopback interface
for this, requiring the portmapper to be run in an insecure
mode.

This patch changes the xdr code such that the request is made to
the loopback interface.

Unfortunately this is not enough to use chimera with rpcbind
without the -i setting (insecure mode). That's because contrary
to the manpage of rpcbind, -i also controls whether the request
has to come from a priviledged port. Without -i only calls
from ports under 1024 are allowed to set an unset.

Acked-by: Gerd Behrmann
Acked-by: Paul Millar
Patch: http://rb.dcache.org/r/4152/