broadcast, cns, hopping, topo, dir, transfermanagers, srm-loginbroker: Define cell.name
The cell.name property is used by our scripts to identify service names
(a better name may be service.name, but we can change that another day).
If cell.name is not defined for a service then the scripts do not
recognize it and there is no way to refer to a particular instance of a
service.

This patch defines cell.name for a number of services for which this
property was previously undefined.

Release notes: Either have a section about the 'dcache' script and
describe that 'dcache services' now lists the services mentioned above,
or have a section for each of those services and describe that they are
now listed by 'dcache services'.

Target: trunk
Request: 2.2
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4472/