all: Check that pid file can be created before starting dcache
The daemon script would start dCache even if the PID file cannot be created
(for whatever reason). The consequence would be that we have a running dCache,
but the scripts would not show it as running and would be unable to stop it.

This patch solves the problem by touching the PID file before starting dCache.
Should that fail then the script will exit.

Release notes: The details are not relevant. A generic comment about improving
the robustness of starting and stopping dCache suffices.

Target: trunk
Request: 2.2
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4475/