all: Avoid error when hostkey is not readable to user
The dcache script checks the format of the hostkey whenever it is invoked.
Since read-access to the hostkey should be restricted, the dcache script will
generate an error whenever the script is invoked by somebody without
read-access to the hostkey.

This patch resolves the problem.

(there is a related question about whether the check should be run for
commands like 'dcache status', but we can discuss that another day)

Release notes: In a section about the dcache script, describe that we 
fixed a bug that caused 'grep: /etc/grid-security/hostkey.pem: Permission 
denied' to be printed.

Target: trunk
Request: 2.2
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4476/