srm: Fix an off-by-one error in directory creation code
The SRM has code to create a directory on upload. The code contains an
off-by-one error. The error makes it impossible to upload a file to the
root of the file system.

This problem has not been discovered before because all sites have a
/pnfs directory in which all files are stored. There is however no
requirement that a dCache instance must have such a directory and
uploading a file into the root directory is perfectly legal.

Release notes: Fixed an issue preventing upload through SRM to the 
root-directory of the name space.

Target: trunk
Request: 2.2
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/4483/