srm: Fix bugs in SRM SQL DDL generation
While working with databases different than postgresql I ran into
situations in which index creation would fail. The failure was to be
expected because the particular database doesn't support that type of
index. However the SRM left behind a long stack trace rather than just
an error message.

The SRM also contains a bug in the code to check whether a table already
exists. The code expects all tables to be lower-case even though some
tables created by the SRM are in upper case. For postgresql this doesn't
make a difference as it translates all identifiers to lower case (unless
quoted), however for other databases this bug causes the SRM to attempt
to create a table that already exists.

Release notes: Does not need to be mentioned.

Target: trunk
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/4485/