srm: Quote count field in SQL
Some databases (hsqldb) consider 'count' to be a reserved word. This
cases table creation to fail in the SRM.

This patch resolves the problem by quoting the field. One challenge is
that quoting the field will cause some databases to preserve the case of
the field. I believe it is reasonable to assume that all existing SRMs
run on postgresql. Postgresql converts unquoted fields to lower case.
For this reason the patch uses lower case for the count field. For other
databases this patch may break the SRM on upgrade.

Release notes: In the section about incompatibilities you should
describe that upgrade of the SRM may fail on databases other than
PostgreSQL. In such cases the SRM 'lsrequests' table has to be dropped.
In a section about the SRM you can describe that compatibility with
databases other than postgresql has been improved.

Target: trunk
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/4487/