srm: Use DataNucleus rather than toplink
The AuthorizationRecord code still used toplink as its JPA
implementation. When experimenting with H2 and HSQLDB I ran into issues
with toplink.

This patch switches to using DataNucleus rather than Toplink. We already
used DataNucleus for the JDO code.

I had to make a few changes to make the code work with DataNucleus.

Unfortunately, I was unable to make DataNucleus map to the same schema
as toplink. The issue is with the boolean field in AuthorizationRecord.
I can work around the problem by forcing a JDBC type of BIT (which works
with PostgreSQL), however then it will not work with HSQLDB and H2.
Forcing a JDBC type of BOOLEAN does not work with PostgreSQL. In the end
I decided to let DataNucleus create new tables. The data in the
authorization tables is not important to preserve. DataNucleus will
recreate the tables using upper case table names and those will be
different from the lower case names used by toplink.

Release notes: Under a section about the SRM you should note that the
authrecord, authgroup and authgrouplist tables of the SRM database
will be replaced by new tables called AUTHRECORD, AUTHGROUP and
AUTHGROUPLIST. Administrators may choose to drop the old tables.

Target: trunk
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/4488/