http: add support for Content-Disposition header in pool response
Web clients that make a GET request normally save the data using the 
last path-element from the URI.  I found some clients would use the 
last path-element of a redirected URI, but also include the query-part 
("?dcache-http-uuid=...").

The 'Content-Disposition' header provides a mechanism to allow the 
web-server to suggest a filename to use when saving data to disk.  
This patch adds support for this header in the reply from the pool.

Since the 'Content-Disposition' head can also need to percent-encode 
the filename (the support of which is buggy in Java and we have a 
work-around), I've refactored the code as a common utility method in a 
string-encoding class.  This simplifies the UrlPathWrapper class, 
which handles this markup for our usage of string-template in the 
WebDAV door.  Unit-tests are included for this new class.

FOR RELEASE NOTES:

With this patch, when a user requests the complete contents of a file 
from dCache and the HTTP/WebDAV door redirects the user to a pool with 
the file, the pool will respond with the Content-Disposition header.  
This header provides a hint to the web-client under what filename 
the data should be stored.

Target: trunk
Patch: http://rb.dcache.org/r/4506/
Acked-by: Gerd Behrmann