billing: Fix billing db properties issue with httpd service
The recent change to use the proper db.* familiy of properties for the
billing database had an unexpected side effect (see
http://rb.dcache.org/r/4473/).

It turns out the billing database breaks the general rule of a database
only being used by a single service (chimera breaks it too, although in
a slightly different way). For the billing database the httpd service
and the billing service have to share the same database instance.

Since the httpd service can embed many different "servlets", it isn't
strictly bound to any particular database, yet it needs access to the
billing database settings. Hence the db.* properties will not work for
the billing database.

The concrete symtom is that the httpd service fails to start when
billing db is enabled. The error message is that db.name is undefined.
The workaround is to restore the original billingDbUrl definition. The
consequence is that redefining db.name and db.host for billing service
will not have the intended effect. I have marked those two properties as
immutable.

Release notes: The patch fixes a problem introduced by another patch
in this release cycle - there is no need for additional comments.

Target: trunk
Request: 2.2
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4507/