chimera-cli: Close connection pool on shutdown
With HSQLDB it is essential that the connections to the database are
closed before the application shuts down. Otherwise changes will not be
flushed to disk.

For test purposes with dCache lost updates may be acceptable, however
for chimera-cli the lifetime is so short that unless the connection gets
properly closed, the cli commands will have absolutely no effect with
HSQLDB.

This patch adds the missing calls to close. Unfortunately I had to do an
instanceof test to do a safe cast to the BoneCP DataSource. The next
release of BoneCP will implement Closeable, allowing us to make the code
a bit cleaner.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4496/