replica: Fix DB connection handling
Replica manager has code to keep a connection to
the database open. This code is back from a time before we had a proper
connection pool.

The code doesn't work in its present form. This is because we after each
statement close the connection no matter whether we use the connection
created in the constructor or a connection created for the operation.
Thus if replica manager were to open a connection in the constructor,
replica manager would quickly fail with errors stating that the
connection had already been closed.

This patch removes the legacy code.

I request this code to be merged to 2.2 to make merging the other fixes
for replica manager easier. This patch in itself has no effect.

Release notes: There is no visible change from this patch. It is enough
to describe the other replica manager patches.

Target: trunk
Request: 2.2
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4500/